Allow determining static concurrency without state
This commit changes the "receiver" of `State::concurrency` from `&Self`
to `Option<&Self>`. That means that the level of concurrency may be
queried even when no state is (locally) available. This is useful when
the level of concurrency never changes, e.g. is always `1`. In such
cases a node may transition out of passive and into active participation
even when it has no state.