Have the default node type be `Send`
The node type returned from `NodeBuilder::build()` was `Shell<N>` and is
now `Shell<N, Sendable>`. As the type argument implies, values of this
type are `Send` and can be turned into `!Send` values of type `Shell<N,
Unsendable>`.

This split into two types is unusual and necessitated by the chosen
implementation strategy. Internally, paxakos uses a [sendable,
non-atomically reference counted smart pointer][snarc]. The trade-off is
that the `Sendable` node's API is slightly awkward. It does not directly
implement `Node`, but only a subset of the trait's methods. To get
access to the full `Node` API, one of its `enter`-flavored method's must
be called.

Because the implementation's artifacts are a bit awkward, it may be
revisited. What ultimately makes the current approach seem acceptable is
that the `Sendable` node's reduced API is generally all that is needed,
everything else being accomplished through `NodeHandle`s.

[snarc]: https://crates.io/crates/snarc